+++ lib/action/attachable/subactions.rb
@@ -5,11 +5,18 @@ module Action
     module Subactions
       extend ActiveSupport::Concern
 
+      included do
+        class_attribute :_axnable_methods, default: {}
+        class_attribute :_axns, default: {}
+      end
+
       class_methods do
         def axnable_method(name, axn_klass = nil, **action_kwargs, &block)
           raise ArgumentError, "Unable to attach Axn -- '#{name}' is already taken" if respond_to?(name)
 
-          action_kwargs[:expose_return_as] ||= :value
+          self._axnable_methods = _axnable_methods.merge(name => { axn_klass:, action_kwargs:, block: })
+
+          action_kwargs[:expose_return_as] ||= :value unless axn_klass
           axn_klass = axn_for_attachment(name:, axn_klass:, **action_kwargs, &block)
 
           define_singleton_method("#{name}_axn") do |**kwargs|
@@ -25,6 +32,8 @@ module Action
         def axn(name, axn_klass = nil, **action_kwargs, &block)
           raise ArgumentError, "Unable to attach Axn -- '#{name}' is already taken" if respond_to?(name)
 
+          self._axns = _axns.merge(name => { axn_klass:, action_kwargs:, block: })
+
           axn_klass = axn_for_attachment(name:, axn_klass:, **action_kwargs, &block)
 
           define_singleton_method(name) do |**kwargs|
@@ -35,6 +44,24 @@ module Action
 
           define_singleton_method("#{name}!") do |**kwargs|
             axn_klass.call!(**kwargs)
+          end
+
+          self._axns = _axns.merge(name => axn_klass)
+        end
+
+        def inherited(subclass)
+          super
+
+          return unless subclass.name.present? # TODO: not sure why..
+
+          # Need to redefine the axnable methods on the subclass to ensure they properly reference the subclass's
+          # helper method definitions and not the superclass's.
+          _axnable_methods.each do |name, config|
+            subclass.axnable_method(name, config[:axn_klass], **config[:action_kwargs], &config[:block])
+          end
+
+          _axns.each do |name, config|
+            subclass.axn(name, config[:axn_klass], **config[:action_kwargs], &config[:block])
           end
         end
       end